# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Finsmart API",
        "description": "API REST para aplicaciones financieras con autenticación",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor de desarrollo"
        }
    ],
    "paths": {
...
        "/api/noticias": {
            "get": {
                "tags": [
                    "Noticias"
                ],
                "summary": "Listar todas las noticias",
                "description": "Obtener lista paginada de noticias publicadas",
                "operationId": "6ddae168a9197d8ea5be0fc4760517d2",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Elementos por página (máximo 50)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    },
                    {
                        "name": "categoria",
                        "in": "query",
                        "description": "Filtrar por categoría",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "tecnologia"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de noticias",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "titulo": {
                                                        "type": "string",
                                                        "example": "Título de la noticia"
                                                    },
                                                    "contenido": {
                                                        "type": "string",
                                                        "example": "Contenido de la noticia..."
                                                    },
                                                    "autor": {
                                                        "type": "string",
                                                        "example": "Juan Pérez"
                                                    },
                                                    "categoria": {
                                                        "type": "string",
                                                        "example": "tecnologia"
                                                    },
                                                    "imagen_url": {
                                                        "type": "string",
                                                        "example": "https://example.com/image.jpg"
                                                    },
                                                    "fecha_publicacion": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-01-15T10:30:00Z"
                                                    },
                                                    "publicado": {
                                                        "type": "boolean",
                                                        "example": true
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Noticias"
                ],
                "summary": "Crear nueva noticia",
                "description": "Crear una nueva noticia (requiere autenticación)",
                "operationId": "70d8ebcd3fb5cd45c7d1c30b818e2ef2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "titulo",
                                    "contenido",
                                    "autor",
                                    "categoria"
                                ],
                                "properties": {
                                    "titulo": {
                                        "type": "string",
                                        "example": "Nueva noticia importante"
                                    },
                                    "contenido": {
                                        "type": "string",
                                        "example": "Contenido completo de la noticia..."
                                    },
                                    "autor": {
                                        "type": "string",
                                        "example": "Juan Pérez"
                                    },
                                    "categoria": {
                                        "type": "string",
                                        "example": "tecnologia"
                                    },
                                    "imagen_url": {
                                        "type": "string",
                                        "example": "https://example.com/image.jpg"
                                    },
                                    "fecha_publicacion": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-01-15T10:30:00Z"
                                    },
                                    "publicado": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Noticia creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "titulo": {
                                            "type": "string",
                                            "example": "Nueva noticia importante"
                                        },
                                        "contenido": {
                                            "type": "string",
                                            "example": "Contenido completo de la noticia..."
                                        },
                                        "autor": {
                                            "type": "string",
                                            "example": "Juan Pérez"
                                        },
                                        "categoria": {
                                            "type": "string",
                                            "example": "tecnologia"
                                        },
                                        "imagen_url": {
                                            "type": "string",
                                            "example": "https://example.com/image.jpg"
                                        },
                                        "fecha_publicacion": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-01-15T10:30:00Z"
                                        },
                                        "publicado": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errores de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}